Fix bug in pypeit_obslog: missing entry_point()
In the recent change to pypeit_obslog to define the entry point and deprecate
the wrapper script (commit 767a1795f1012dae4288a779b5917adefd24d244), the
actual entry_point() function was not defined in obslog.py.

This commit adds the function entry_point() to match other scripts in this
directory.

	modified:   pypeit/scripts/obslog.py