Addressing PR comments: `import *` / test for directory existance
Add `# noqa` to `import *`, and also add `__all__` to limit what is imported.

Moved directory existance tests into `Paths` class.  This necessitated the use
of a metaclass -- at least until PypeIt advances to python>=3.9.  At that time,
we can combine `@classmethod` with `@property` for the property methods that do
the existance checking.

	modified:   pypeit/data/__init__.py
	modified:   pypeit/data/utils.py
	modified:   pypeit/scripts/install_ql_masters.py
	modified:   pypeit/scripts/install_telluric.py