Save manual wavelength identification to cache; Doc changes
Save the output of `pypeit_identify` to the PypeIt cache for direct use.  This
includes small modifications to the `pypeit_identify` script to pass down
appropriate information for the ad hoc filename.

Update the `pypeit_install_telluric` script to download the TelGrid files into
the cache.

Update the "Additional Data" and "Wavelength Calibration" documentation to
reflect the changes in this commit.

	modified:   doc/installing.rst
	modified:   doc/wave_calib.rst
	modified:   pypeit/core/gui/identify.py
	modified:   pypeit/data/utils.py
	modified:   pypeit/par/pypeitpar.py
	modified:   pypeit/scripts/identify.py
	modified:   pypeit/scripts/install_telluric.py