Maintain cache name for TelGrid files even if S3 URL changes
Utilize the `sources` feature of AstroPy's `download_file()` to maintain the
cache name (encoded as the source URL) for Tellurig Grid files while pointing
to updated URLs, should the need arise.  This will prevent users from having
to re-download TelGrid files (4-8GB) just because the S3 server URL changed,
because the cache is tied to the source URL.  The cache name is now enshrined
(carved in stone, as it were) with the current (2022-04-24) S3 URL, but that
is hashed and the user won't be bothered with it.

If changes in the S3 URL are required, simply update the URL in
`pypeit/data/s3_url.txt` on the `release` branch, and the code will point to the
correct file, caching it under the old URL name.

	modified:   pypeit/data/utils.py
	modified:   pypeit/tests/test_pkgdata.py