Able to run pypeit_setup (with some bugs)
Shifted all the header metadata with fixed values (ie if 'dispname' which is the name of the dispersing element is always the same / doesn't exist in the header) from the function that maps the header names to the pypeit-specific metadata keywords to the function that defines the values of cards that either don't exist in the header OR exist such that we cannot just hand it over to Pypeit as-is.

Fixed (to an extent) the query that asks "Is this frame an arc?", to where the pypeit_setup command will RUN; however, the command cannot actually IDENTIFY any arcs, which is the whole purpose of the "Is this frame an arc?" query. Alas. The check_frrame_type function (or whatever it is called) cannot identify flats, either.

SO next time, go through the raw data's headers and change all the ones with 'OBJECT' = '....... Arc' to 'IMAGETYP' = 'Comp         '
(Aka the shift arcs)
Then see how to better define the flats