Add option to `pypeit_identify` to allow residual plot to auto-rescale
Especially when performing new manual wavelength IDs for echelle spectrographs,
the residuals can become quite large.  By allowing the residual plot to auto-
rescale, outlier identifications can be more easily marked and dealt with.

This commit does not change the default behavior of `pypeit_identify`, and the
use of the new flag is reuqired to invoke the new functionality.

	modified:   pypeit/core/gui/identify.py
	modified:   pypeit/scripts/identify.py