Change meaning of `excessive_shift = skip` parameter
Setting `excessive_shift` to `skip` now skips 1D spectral extraction, as if
there were no objects found in the frame.

This change was made because it is unlikely that a shift of zero pixels is
correct, and simply skipping extraction has the effect of removing the
problematic frame from the final list of 1D spectra.

	modified:   doc/flexure.rst
	modified:   doc/pypeit_par.rst
	modified:   pypeit/core/flexure.py
	modified:   pypeit/extraction.py
	modified:   pypeit/par/pypeitpar.py