Add `verbosity` option to post-processing scripts
The default verbosity level for `run_pypeit` is 2 (with an option to change it),
but the other scripts have a default verbosity = 1 with no means to change it
from the command line.  Given that several scripts do heavy post-processing
(coadding, flux calibrating, etc.), debugging of user issues can be more
challenging without the higher verbosity level (and logging).

This change adds the `--verbosity N` option to various post-processing scripts.
In addition to being able to change the verbosity, when set to Level 2 a log
file is also produced.  The log filename includes the name of the script and a
timestamp (to the minute) for reference in the event of multiple runs of the
script.

	modified:   pypeit/scripts/coadd_1dspec.py
	modified:   pypeit/scripts/coadd_2dspec.py
	modified:   pypeit/scripts/coadd_datacube.py
	modified:   pypeit/scripts/collate_1d.py
	modified:   pypeit/scripts/flux_calib.py
	modified:   pypeit/scripts/sensfunc.py
	modified:   pypeit/scripts/skysub_regions.py
	modified:   pypeit/scripts/tellfit.py
	modified:   pypeit/scripts/trace_edges.py
	modified:   pypeit/scripts/utils.py