Include S/N ratio for extracted spectra in the ``SpecObj`` datamodel
The median signal-to-noise ratio for extracted spectra are used in various
places (flexure, collation) and also computed for the 1D txt output table, but
are not explicitly included in the ``SpecObj`` datamodel (and hence included in
the spec1d FITS headers).

This commit adds the computed S/N ratio to that data model and also modifies
the users of the value to point to the datamodel value rather than recompute it.

	modified:   pypeit/core/flexure.py
	modified:   pypeit/extraction.py
	modified:   pypeit/scripts/collate_1d.py
	modified:   pypeit/specobj.py
	modified:   pypeit/specobjs.py