Align docs/params to code reality in overscan poly
The code for fitting a polynomial to the overscan region of a raw
image file for subtraction only requires one parameter (order), whereas
the documentation and PypeIt parameter fed into the overscan-subtraction
routine say "three".  This commit aligns the documentation and params
with the reality of the code.  Also, for the polynomial fitting, move
away from the `np.polyfit()` function (deprecated in NumPy v1.4 -- way
way long time ago) to the `numpy.polynomial` pagkage.  In particular,
use a Chebyshev polynomial for the fitting here (orthonormal basis sets
can be useful for seeing where the power is).

Modify the LDT/DeVeny params to use a polynomial (order=1) for fitting
the overscan.  This is more appropriate for this chip and EMI pickup
environment than the default SavGol filter.

	modified:   pypeit/core/procimg.py
	modified:   pypeit/par/pypeitpar.py
	modified:   pypeit/spectrographs/ldt_deveny.py