HOTFIX: Incorrect platescale binning correction
In `pypeit.find_objects.MultiSlitFindObjects.get_platescale()`, the
platescale is scaled by the binning -- but the code accidentally was
using the *spectral* binning rather than the *spatial* binning.  This
value is used to compute the BOXCAR RADIUS in the
`pypeit.core.findobj_skymask.objs_in_slit()` function, which is then
saved in the `SpecObj` object.

This commit corrects the binning scaling to use the SPATIAL binning.

	modified:   pypeit/find_objects.py