Add compound meta that reads SCITARG -> 'target'
For long target lists, some observers do not type the individual object
names into the LOUI, but rather rely upon the SCITARG keyword that comes
from the TCS via the ObserverTargetList tool.  This results in the
OBJECT/OBJNAME keywords being set to UNKNOWN.  When PypeIt tries to read
in the metadata, this leaves all object frames with the same, unhelpful
name.  To options are possible for dealing with this use case.

Option #1: Add a compound meta case for 'target' that uses the SCITARG
FITS keyword for OBJECT frames if the OBJNAME is "UNKNOWN" or blank.
This means that only science frames could have the 'target' metadata
pulled from SCITARG, leaving all calibration frames as before.

	modified:   doc/releases/1.14.1dev.rst
	modified:   pypeit/spectrographs/ldt_deveny.py